experiment/source/ai-app/requirements_streamlit.txt
streamlit>=1.28.0
opencv-python>=4.8.0
ultralytics>=8.0.0
pyttsx3>=2.90
pygame>=2.5.0
numpy>=1.24.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0

